refactor UI Keyboard Event

+++ src/core/src/sge_core/native_ui/win32/Win32Util.h
@@ -4,7 +4,7 @@
 
 namespace sge {
 
-enum class UIKeyboardEventKeyCode : u64;
+enum class UIKeyboardEventKeyCode : u32;
 
 struct Win32Util {
 	using KeyCode = UIKeyboardEventKeyCode;
